Export lazy player from its component
